Kefka SVN Update: Lots of background work that is little fun to ya'll. :D
Caster/Ranged classes have had an override done on them for Zekken and Deep Dungeon specifically for dodging and interrupts. The latter no longer is working through Kefka. Please be aware that this is a temporary band-aid type fix, and I have no idea how it is going to affect the upper levels of the classes, but, it'll allow for dodging at least to take place. Good luck!